move common code to shared object
